Introduce plotting features, only rank on active players.
To make the ranker better, we introduce a feature where we can plot the chart of
energy. Once we have a plot, we can see the ranker is failing to produce good
rankings and is going for a corner of the solution space.

By altering the ranking criterion to look at active players only, we get much better
ranking results compared to earlier. Say you are looking at tournaments 0-152. Then
we find active players in tournaments 150, 151 and 152 and use prediction from
tournaments 147, 148 and 149. We ranked on 0-146 in order to predict.

The reason is works so much better is that stray players are not being considered
anymore.